DATAGRAPH-582 - Cannot replace an existing RelationshipEntity.
Refactored.
