+++ neo4j-ogm/src/main/java/org/neo4j/ogm/cypher/compiler/CypherCompiler.java
@@ -43,6 +43,7 @@ public interface CypherCompiler {
      * @param startNode The {@link NodeBuilder} representation of the relationship start node
      * @param relationshipType The type of relationship between the nodes to delete
      * @param endNode The {@link NodeBuilder} representation of the relationship end node
+     * @param relId The id of the relationship to unrelate
      */
     void unrelate(String startNode, String relationshipType, String endNode, Long relId);
