+++ neo4j-ogm/src/main/java/org/neo4j/ogm/mapper/GraphEntityMapper.java
@@ -187,9 +187,7 @@ public class GraphEntityMapper implements GraphToEntityMapper<GraphModel> {
 
                     if (writer.forScalar()) {
                         writer.write(source, relationshipEntity);
-                        MappedRelationship mappedRelationship = new MappedRelationship(edge.getStartNode(), edge.getType(), edge.getEndNode());
-                        mappedRelationship.setRelId(edge.getId());
-                        mappingContext.registerRelationship(mappedRelationship);
+                        mappingContext.registerRelationship(new MappedRelationship(edge.getStartNode(), edge.getType(), edge.getEndNode(), edge.getId()));
                     } else {
                         oneToMany.add(edge);
                     }
@@ -222,9 +220,7 @@ public class GraphEntityMapper implements GraphToEntityMapper<GraphModel> {
                 if (!oneToOne) {
                     oneToMany.add(edge);
                 } else {
-                    MappedRelationship mappedRelationship =  new MappedRelationship(edge.getStartNode(), edge.getType(), edge.getEndNode());
-                    mappedRelationship.setRelId(edge.getId());
-                    mappingContext.registerRelationship(mappedRelationship);
+                    mappingContext.registerRelationship(new MappedRelationship(edge.getStartNode(), edge.getType(), edge.getEndNode(), edge.getId()));
                 }
             }
         }
@@ -304,9 +300,7 @@ public class GraphEntityMapper implements GraphToEntityMapper<GraphModel> {
 
         // finally register all the relationships in the mapping context
         for (RelationshipModel edge : oneToManyRelationships) {
-            MappedRelationship mappedRelationship = new MappedRelationship(edge.getStartNode(), edge.getType(), edge.getEndNode());
-            mappedRelationship.setRelId(edge.getId());
-            mappingContext.registerRelationship(mappedRelationship);
+            mappingContext.registerRelationship(new MappedRelationship(edge.getStartNode(), edge.getType(), edge.getEndNode(), edge.getId()));
         }
     }
