+++ neo4j-ogm/src/main/java/org/neo4j/ogm/mapper/TransientRelationship.java
@@ -32,6 +32,7 @@ import java.util.Map;
  *
  * @author Mark Angrish
  * @author Vince Bickers
+ * @author Luanne Misquitta
  */
 public class TransientRelationship {
 
@@ -57,16 +58,7 @@ public class TransientRelationship {
 
         Long srcIdentity = src.startsWith("_") ? refMap.get(src) : Long.parseLong(src.substring(1));
         Long tgtIdentity = tgt.startsWith("_") ? refMap.get(tgt) : Long.parseLong(tgt.substring(1));
-        Long relIdentity=null;
-        if(ref != null) {
-            if(ref.startsWith("_")) {
-                relIdentity = refMap.get(ref);
-            }
-            else {
-                relIdentity = Long.parseLong(ref.substring(1));
-            }
-        }
-        //Long relIdentity = ref == null ? null : ref.startsWith("_") ? refMap.get(ref) : Long.parseLong(ref.substring(1));
+        Long relIdentity = ref.startsWith("_") ? refMap.get(ref) : Long.parseLong(ref.substring(1));
 
         if (srcIdentity == null) {
             throw new RuntimeException("Couldn't get identity for " + src);
@@ -76,9 +68,11 @@ public class TransientRelationship {
             throw new RuntimeException("Couldn't get identity for " + tgt);
         }
 
-        MappedRelationship mappedRelationship = new MappedRelationship(srcIdentity, rel, tgtIdentity);
-        mappedRelationship.setRelId(relIdentity);
-        return mappedRelationship;
+        if (relIdentity == null) {
+            throw new RuntimeException("Couldn't get identity for " + ref);
+        }
+
+        return new MappedRelationship(srcIdentity, rel, tgtIdentity, relIdentity);
     }
 
     public boolean equalsIgnoreDirection(String src, String type, String tgt) {
